mcp/requirements.txt
fastmcp
psycopg2-binary
pydantic
python-dotenv

